Move Gmail Apps integration from prototype code into production ready code
Integrate creating Gmail accounts into deploy method on EventApplication

Add missing requirements for Google Apps integration
