+++ djangogirls/settings.py
@@ -282,3 +282,6 @@ if 'OPBEAT_SECRET_TOKEN' in os.environ:
     ] + MIDDLEWARE
 
 CODEMIRROR_PATH = "vendor/codemirror/"
+
+GOOGLE_APPS_ADMIN_SDK_CREDENTIALS = os.path.join(BASE_DIR, '..', 'djangogirls_website-93134565f52d.json')
+GOOGLE_APPS_ADMIN_SDK_SCOPES = 'https://www.googleapis.com/auth/admin.directory.user'
