+++ requirements.txt
@@ -32,6 +32,8 @@ docopt==0.4.0             # via mandrill
 docutils==0.13.1          # via botocore
 easy_thumbnails==2.3
 freezegun==0.3.8
+google-api-python-client==1.6.0
+httplib2==0.9.2
 icalendar==3.11.2
 jmespath==0.9.0           # via boto3, botocore
 lxml==3.7.2
